Fix setting for TMPPREFIX.
From the docs:
TMPPREFIX  A pathname prefix which the shell will use for all temporary
files.  Note that this should include an initial part for the file name
as well as any directory names.  The default is `/tmp/zsh'.

This was causing weird messages when using conda to switch environments.